data_flow_and_processing/stream_processing/stream-stock/websocket_stream/requirements.txt
kafka-python
fastapi
uvicorn[standard]
websockets